simulateur/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
cachecontrol==0.14.0
certifi==2023.11.17
cffi==1.17.0
channels==4.1.0
channels-redis
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
crashtest==0.4.1
cryptography==43.0.0
cssselect2==0.7.0
cycler==0.12.1
daphne==2.5.0
defusedxml==0.8.0rc2
distlib==0.3.8
django==5.0.8
django-allauth==64.1.0
django-allauth-ui==1.4.1
django-any-urlfield==2.7.post1
django-appconf==1.0.6
django-check-seo==1.0.1
django-classy-tags==4.1.0
django-cms==3.11.6
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-easy-audit==1.3.6
django-extensions==3.2.3
django-filer==3.1.4
django-formtools==2.5.1
django-guest-user==0.5.5
django-health-check==3.18.3
django-imagekit==5.0.0
django-jet-reboot==1.3.7
django-meta==2.4.2
django-money==3.5.3
django-organizations==2.4.1
django-perf-rec==4.26.0
django-pictures==1.3.3
django-polymorphic==3.1.0
django-prometheus==2.3.1
django-pwa==2.0.0
django-robots==6.1
django-rq==2.10.2
django-sekizai==4.1.0
django-sslserver==0.22
django-storages==1.12.3
django-tinymce==4.1.0
django-treebeard==4.7.1
django-typer==2.2.1
django-webpack-loader==3.1.0
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
djangocms-page-meta==1.4.4
djangorestframework==3.15.2
drf-yasg==1.21.7
dulwich==0.21.7
easy-thumbnails==2.9
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.15.4
fonttools==4.53.1
google-api-python-client==1.4.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
installer==0.7.0
jaraco-classes==3.4.0
jinja2==3.1.4
keyring==24.3.1
kiwisolver==1.4.5
kombu==5.4.0
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.4.0
noise==1.2.2
numpy==2.1.0
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.3.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pilkit==3.0
pillow==10.4.0
pip==24.2
pkginfo==1.11.1
platformdirs==4.2.2
poetry
prometheus-client==0.20.0
prompt-toolkit==3.0.47
psutil==6.0.0
psycopg==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py-moneyed==3.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
pygments==2.18.0
pyjwt==2.9.0
pyopenssl==24.2.1
pyparsing==3.1.2
pyproject-hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pywin32-ctypes==0.2.3
pyyaml==6.0.2
rapidfuzz==3.9.6
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rq==1.16.2
rsa==4.9
service-identity==24.1.0
setuptools==72.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
slippers==0.6.2
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tinycss2==1.3.0
tomli==2.0.1
tomlkit==0.13.2
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
trove-classifiers==2024.7.2
twisted==24.7.0
txaio==23.1.1
typeguard==2.13.3
typer-slim==0.12.4
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uv==0.2.37
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
websockets-cli==0.3.0
wheel==0.44.0
whitenoise==6.7.0
wsproto==1.2.0
zope-interface==7.0.1
twisted[tls,http2]